init basic test case for scaling
